Fixed Delphi 2009 compilation;  still some conversion warnings, and everything still needs to be audited.
Minor stylistic cleanup;  lots of work remaining.
Fixed SetLinkFlag supported (necessary when creating symlinks).
Imported a couple of minor changes from Cobines.
Archive tail (1024*0x00) is only written if the new archive has at least one item in it.  This is different than Cobines, and open to change.
Fixed SaveArchive so it creates a new TFileStream instead of a TAbSpanStream.