Tightened up TAbListView/TAbTreeView communication and property setting.
Setting the listview's TreeView property or the treeview's ListView
property will now set the opposite property for the other component,
and will automatically set the Archive property.

Improved removal notification so if only one of the components is
destroyed the other will take over OnChange events.

If the treeview's Path property is set before the Archive property it will
still select the path node, and it will reselect it if the archive is
changed.

git-svn-id: http://svn.code.sf.net/p/tpabbrevia/code/trunk@440 89253570-fd12-4709-a508-8497ac4b1c7b