Fixed crash when extracting a gzip file using TAbUnzipper.
Fixed InitArchive in the component classes using either "FArchive <> nil" or "FArchive is TAbZipArchive" incorrectly, so either the class was being cast to a TAbZipArchive incorrectly or properties weren't being set if the archive wasn't a zip.

Replaced all internal usage of the "ZipArchive" property with "FArchive" since most of the functions aren't actually dependent on the archive class.  ZipArchive is actually a TAbArchive, but the name is unnecessarily confusing.

git-svn-id: http://svn.code.sf.net/p/tpabbrevia/code/trunk@514 89253570-fd12-4709-a508-8497ac4b1c7b