[ 3542662 ] Fixed TAbVirtualMemoryStream support for files larger than 2GB on older Delphi releases.
TAbVirtualMemoryStream.vmsSwapFileWrite() was calling FileSeek() and at least on Delphi 7 that was calling the overload for a 32-bit offset rather than a 64-bit one..  Adding a cast around the vmsSwapHandle parameter fixed the issue, but the cast would have had to be different in later Delphi releases since the Handle parameter changed from a LongInt to a THandle.  Cleaner to just use a TFileStream everywhere.