Fix: TAbZipper.AddFiles() is not thread safe.
When calling TAbZipper.AddFiles() from separate instances of TAbZipper from separate threads the process of zipping the files can fail. This is because of the internal calls to ChDir that are used through the source code. Usages of ChDir() can be found in: AbZipPrc.pas, AbTarTyp.pas, AbGzTyp.pas, AbArcTyp.pas.
The problem occurs because Thread #1 will come change the current working directory with a call to ChDir() and if Thread #2 comes and does the same thing right after thread #1 does, Thread #1 will now be operating on a directory that it is not supposed to be working on because thread #2 changed the working directory before thread #1 had a chance to do any work.

Symptoms to the problem will be that the application will have directories locked after the two threads get done doing their work.

This was fixed by removing the calls to ChDir() and keeping track of the current directory being operated on.